Add BMP redundant connection detection logic

+++ bgp/bmp-impl/src/main/java/org/opendaylight/protocol/bmp/impl/app/BmpRouterImpl.java
@@ -83,20 +83,31 @@ public class BmpRouterImpl implements BmpRouter, TransactionChainListener {
     }
 
     @Override
-    public void onSessionUp(final BmpSession session) {
+    public boolean onSessionUp(final BmpSession session) {
         this.session = session;
         this.routerIp = InetAddresses.toAddrString(session.getRemoteAddress());
         this.routerId = new RouterId(Ipv4Util.getIpAddress(session.getRemoteAddress()));
+        // check if this session is redundant
+        if (sessionManager.isSessionExist(this)) {
+            LOG.warn("Redundant session (local <-> {}) detected. Connection will be abandoned.", session.getRemoteAddress().getHostAddress());
+            this.routerId = null;
+            return false;
+        }
         this.routerYangIId = YangInstanceIdentifier.builder(this.sessionManager.getRoutersYangIId()).nodeWithKey(Router.QNAME,
                 ROUTER_ID_QNAME, this.routerIp).build();
         this.peersYangIId = YangInstanceIdentifier.builder(routerYangIId).node(Peer.QNAME).build();
         createRouterEntry();
         this.sessionManager.addSessionListener(this);
+        return true;
     }
 
     @Override
     public void onSessionDown(final BmpSession session, final Exception e) {
+        // turn down session only when the current session is valid
+        if (!isSessionValid())
+            return;
         LOG.info("Session {} went down.", session);
+        LOG.debug("Exception found", e);
         tearDown();
     }
 
@@ -136,15 +147,21 @@ public class BmpRouterImpl implements BmpRouter, TransactionChainListener {
         } catch(final Exception e) {
             LOG.error("Failed to properly close BMP application.", e);
         } finally {
-            try {
-                final DOMDataWriteTransaction wTx = this.domDataBroker.newWriteOnlyTransaction();
-                wTx.delete(LogicalDatastoreType.OPERATIONAL, this.routerYangIId);
-                wTx.submit().checkedGet();
-            } catch (final TransactionCommitFailedException e) {
-                LOG.error("Failed to remove BMP router data from DS.", e);
+            // remove session only when session is valid, otherwise
+            // we would remove the original valid session when a redundant connection happens
+            // as the routerId is the same for both connection
+            if (isSessionValid()) {
+                try {
+                    // it means the session was closed before it was written to datastore
+                    final DOMDataWriteTransaction wTx = this.domDataBroker.newWriteOnlyTransaction();
+                    wTx.delete(LogicalDatastoreType.OPERATIONAL, this.routerYangIId);
+                    wTx.submit().checkedGet();
+                } catch (final TransactionCommitFailedException e) {
+                    LOG.error("Failed to remove BMP router data from DS.", e);
+                }
+                this.sessionManager.removeSessionListener(this);
             }
         }
-        this.sessionManager.removeSessionListener(this);
     }
 
     @Override
@@ -154,10 +171,15 @@ public class BmpRouterImpl implements BmpRouter, TransactionChainListener {
 
     @Override
     public void onTransactionChainSuccessful(final TransactionChain<?, ?> chain) {
-        LOG.debug("Transaction chain {} successfull.", chain);
+        LOG.debug("Transaction chain {} successfully.", chain);
+    }
+
+    private boolean isSessionValid() {
+        return !(this.routerId == null || this.routerYangIId == null);
     }
 
     private void createRouterEntry() {
+        Preconditions.checkArgument(isSessionValid());
         final DOMDataWriteTransaction wTx = this.domTxChain.newWriteOnlyTransaction();
         wTx.put(LogicalDatastoreType.OPERATIONAL, this.routerYangIId,
                 Builders.mapEntryBuilder()
@@ -169,6 +191,7 @@ public class BmpRouterImpl implements BmpRouter, TransactionChainListener {
     }
 
     private void onInitiate(final InitiationMessage initiation) {
+        Preconditions.checkArgument(isSessionValid());
         final DOMDataWriteTransaction wTx = this.domTxChain.newWriteOnlyTransaction();
         wTx.merge(LogicalDatastoreType.OPERATIONAL, this.routerYangIId,
                 Builders.mapEntryBuilder()
